hw_09/pyproject.toml
[project]
name = "hw-09"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
]
